chore: refactor channel reward manager
